src/main/java/ru/vitaly/_3_LongestSubstringWithoutRepeatingCharacters.java
package ru.vitaly;

import java.util.HashMap;
import java.util.Map;

/**
 * @author Vitaly Vasilyev, e-mail: rav.energ@rambler.ru
 * @version 1.0
 * @since 25.01.2022
 */
public class _3_LongestSubstringWithoutRepeatingCharacters {
    public static int lengthOfLongestSubstring(String s) {
        Map<Character, Integer> map = new HashMap<>();
        int left = 0;
        int subStrLen = 0;
        for (int i = 0; i < s.length(); i++) {
            char c = s.charAt(i);
            if (map.containsKey(c)) {
                left = Math.max(left, map.get(c) + 1);
            }
            map.put(c, i);
            subStrLen = Math.max(subStrLen, i - left + 1);
        }
        return subStrLen;
    }
}